// src/types/mock-data.type.ts
export type MockData = {
  titles: string[];
  descriptions: string[];
  genres: string[];
  filmYears: string[];
  ratings: string[];
  previews: string[];
  videos: string[];
  actors: string[];
  directors: string[];
  filmLengths: string[];
  commentsCounts: string[];
  names: string[];
  emails: string[];
  avatars: string[];
  posters: string[];
  backgroundImages: string[];
  colorBackgroundImages: string[];
};